Fixed page structuring and ither issues
